requirements.txt
certifi==2017.4.17
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.4.1
Django==1.11.1
django-allauth==0.32.0
django-cors-headers==2.1.0
django-filter==1.0.4
djangorestframework==3.6.4
gunicorn==19.6.0
idna==2.5
Markdown==2.6.9
oauthlib==2.0.2
psycopg2==2.6.2
python-http-client==3.0.0
python3-openid==3.1.0
pytz==2017.2
requests==2.18.1
requests-oauthlib==0.8.0
sendgrid-django
urllib3==1.21.1
whitenoise==3.2
